+++ src/Times.ts
@@ -1,5 +1,3 @@
 export function times<T>(n: number, f: (idx: number) => T): T[] {
-  return Array(n)
-    .fill(undefined)
-    .map((_, i) => f(i))
+  return Array.from({ length: n }, (_, i) => f(i))
 }